Do not store local service account token and CA to config.
When defaulting to local JWT token and CA certificate in a pod, always read
them from local filesystem and do not store them persistenly with the config.

Token will be re-read periodically to avoid using expired token.

The change allows running Vault on Kubernetes 1.21 and newer, which switched
to ID token that is bound to the pod and will expire.

Signed-off-by: Tero Saarni <tero.saarni@est.tech>