Return default value for role alias_name_source instead of empty string
Field alias_name_source is missing if role was created with Vault prior
version 1.9.0.  When reading stored role back, the go struct entry gets
initialized to "" and user will be presented role with "alias_name_source": ""
in HTTP response, which is invalid value.  This change returns the default
value instead of empty string.

Signed-off-by: Tero Saarni <tero.saarni@est.tech>