Switch from golant-jwt to cap+jose
This caused stricter token checking around for Issued At and Expiration,
so we cleaned up the tokens in the tests to be stored in plain JSON and
signed as part of the test initialization to make it easier to see what
is going on. This also means that the tests won't start mysteriously
failing in 2030. :)

We also updated to the latest `hashicorp/cap` library to support
Kubernetes 1.20 tokens.