Add path support for kubernetes/config
It can be usefull to be able to change the value of some parameters in
the config but not others.

For example when changing disable_iss_validation we needed to keep
token_reviewer_jwt as was already set, but since this value is not
returned by pathConfigRead() there is no way to do this if you do not
already do know what value it was set to.

This patch adds PATCH support to the kubernetes/config endpoint so that
any parameter can be updated while not changing the other ones.