Ensure a consistent TLS configuration
Previously, it was possible for the http.Client's Transport to be
missing the necessary root CAs to ensure that all TLS connections
between the auth engine and the Kubernetes API were validated against a
configured set of CA certificates.

This fix ensures that the http.Client's Transport is always consistent
with the configured CA cert chain, but introducing a periodic TLS
configuration checker that is started as part of the backend's
initialization. The periodic checker provides a signal handler that will
update the backend's TLS configuration on SIGHUP.

Other fixes:
- only update the client's transport when the CA certificate pool has
  changed.