Ensure a consistent TLS configuration (#173) (#178)
* Ensure a consistent TLS configuration for k8s API requests

Previously, it was possible for the http.Client's Transport to be
missing the necessary root CAs to ensure that all TLS connections
between the auth engine and the Kubernetes API were validated against a
configured set of CA certificates.

This fix ensures that the http.Client's Transport is always consistent
with the configured CA cert chain, by introducing a periodic TLS
configuration checker that is started as part of the backend's
initialization.

Other fixes:
- only update the client's transport when the CA certificate pool has
  changed.

Co-authored-by: Tom Proctor <tomhjp@users.noreply.github.com>